move colab method to util
